RISCV: Clear timer interrupt pending bit.
The timer interrupt pending bit is cleared in bbl while
reprogramming the timer. This works fine unless we are
in nohz mode. Since, the timer is not reprogrammed, the
pending bits are not cleared leading to continuous timer
interrupt firing and cpu stalls.

Clear timer interrupt in timer interrupt handler by calling
into SEE. This also introduces an SBI call to do this.

This patch requires following bbl fix.
https://github.com/riscv/riscv-pk/pull/108

Both kernel & bbl patch are required to run kernel as tick less
mode in RISC-V.

The details of the stalls can be found in
https://github.com/riscv/riscv-linux/issues/127

Signed-off-by: Atish Patra <atish.patra@wdc.com>